Ensure test precondition.
As written the tests in test_tools assume the tests will be run in the
root of the source and that a file named setup.py will exist.

The Debian test runner installs the package into a subdirectory,
and then runs the tests there.

This adds a test setUp and tearDown to add (and remove) an empty
setup.py if no setup.py file already exists