migrate from external mock package to stdlib unittest.mock (#212)
The PyPI standalone mock package is a straight backport of the stdlib to
older versions of python. It is usually not needed.

In this case, the required version of mock is >=4, which backports the
python 3.8 stdlib. The minimum version of python required is already
3.8, so all functionality guaranteed to exist is already part of the
stdlib.

Simply use the stdlib directly.